feat: xpra support in Jarvice Desktop
Initial commit from Marc Levrier

+++ nimbix_desktop/postinstall-desktop.sh
@@ -1,18 +1,23 @@
-SKEL_CONFIGS=$(cd $dirname/skel.config && find . -type f)
-for i in $SKEL_CONFIGS; do
+if [ "${XPRA}" = false ]; then
+  SKEL_CONFIGS=$(cd $dirname/skel.config && find . -type f)
+  for i in $SKEL_CONFIGS; do
     mkdir -p $(dirname /etc/skel/.config/$i)
     cp -f $dirname/skel.config/$i /etc/skel/.config/$i
     chmod u+w /etc/skel/.config/$i
-done
+  done
 
-# Copy in the config files to set Firefox as default
-cp -f $dirname/helpers.rc /etc/skel/.config/xfce4
-cp -f $dirname/mimeapps.list /etc/skel/.config
+  # Copy in the config files to set Firefox as default
+  cp -f $dirname/helpers.rc /etc/skel/.config/xfce4
+  cp -f $dirname/mimeapps.list /etc/skel/.config
+fi
 
 rm -f /usr/local/bin/nimbix_desktop
 ln -sf $dirname/nimbix_desktop /usr/local/bin/nimbix_desktop
-rm -f /usr/local/bin/xfce4-session-logout
-ln -sf $dirname/xfce4-session-logout /usr/local/bin/xfce4-session-logout
+
+if [ "${XPRA}" = false ]; then
+  rm -f /usr/local/bin/xfce4-session-logout
+  ln -sf $dirname/xfce4-session-logout /usr/local/bin/xfce4-session-logout
+fi
 
 # VirtualGL is now provided by platform; link in the binaries that will
 # be deployed if VGL is available
@@ -30,5 +35,9 @@ done
 
 mkdir -p /etc/NAE
 if [ ! -e /etc/NAE/url.txt ]; then
-    echo 'https://%PUBLICADDR%:5902/vnc.html?password=%NIMBIXPASSWD%&autoconnect=true&reconnect=true' >/etc/NAE/url.txt
+  if [ "${XPRA}" = true ]; then
+    echo 'https://%PUBLICADDR%:5903/index.html?handle=owner&token=%NIMBIXPASSWD%' >/etc/NAE/url.txt
+  else
+    echo 'https://%PUBLICADDR%:5902/vnc.html?password=%NIMBIXPASSWD%&autoconnect=true&reconnect=true' >/etc/NAE/url.txt 
+  fi
 fi
